+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Added additional inventor benchmark datasets.
+
 ## 0.0.1 (June 27, 2022)
 
 - Initial release